requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
pdfplumber
python-docx
nltk
scikit-learn
reportlab
Jinja2